llvm.sh accepts arguements

+++ llvm.sh
@@ -1,5 +1,28 @@
 #!/bin/bash
 
+function test
+{
+    DIR="$(dirname "$1")"
+    filename="$(basename "$1")"
+
+    echo "Directory: $DIR Filename: $filename"
+
+    name="${filename%.*}"
+
+    if java Main "$DIR"/"$name".mini > "$OUT"/"$name".str
+    then
+        mv "$DIR"/"$name".ll "$OUT"/"$name".ll
+
+        if clang -o "$OUT"/"$name".bin "$OUT"/"$name".ll
+        then
+            if ! "$OUT"/"$name".bin
+            then
+                code -w "$OUT"/"$name".ll ./examples/llvm/"$name".llvm "$DIR"/"$name".mini "$OUT"/"$name".str
+            fi
+        fi
+    fi
+}
+
 DIR="./examples/positive"
 
 OUT="./IR"
@@ -9,22 +32,19 @@ mkdir -p "$OUT"
 ./compile.sh --clean
 ./compile.sh
 
-for file in $(ls "$DIR")
+if [ ! "$#" -eq 0 ]
+then
+    for filename in "$@"
+    do
+        test "$filename"
+    done
+
+    exit 0
+fi
+
+for filename in $(ls "$DIR")
 do
-    name="${file%.*}"
-
-    if java Main "$DIR"/"$file" > "$OUT"/"$name".str
-    then
-        mv "$DIR"/"$name".ll "$OUT"/"$name".ll
-
-        if clang -o "$OUT"/"$name".bin "$OUT"/"$name".ll
-        then
-            if ! "$OUT"/"$name".bin
-            then
-                code -w "$OUT"/"$name".ll ./examples/llvm/"$name".llvm "$DIR"/"$file" "$OUT"/"$name".str
-            fi
-        fi
-    fi
+    test "$filename"
 done
 
 rm -rfv "$OUT"
